[BE] 컨텐츠 도메인 생성 및 코드 이전 (#133)
* feat: contents 앱 생성

* chore: contents 앱 전역 등록

* chore: Account 앱의 home_statistics 코드 이전

* test: home statistics 임시 테스트 코드 작성
