+++ backend/account/views/oj.py
@@ -35,38 +35,10 @@ from ..serializers import (ApplyResetPasswordSerializer, ResetPasswordSerializer
                            DashboardSubmissionSerializer,
                            DashboardDepartmentSerializer, DashboardCollegeSerializer, DashboardRankSerializer,
                            DashboardUserInfoSerializer, DashboardFieldInfoSerializer,
-                           DashboardDifficultyInfoSerializer,
-                           HomeStatistics)
+                           DashboardDifficultyInfoSerializer)
 from ..serializers import (TwoFactorAuthCodeSerializer, UserProfileSerializer,
                            EditUserProfileSerializer, ImageUploadForm)
 from ..tasks import send_email_async
-
-class GetHomeStatisticsAPI(APIView):
-    def get(self, request):
-        """
-        총 문제 수, 채점이 완료된 문제 수, 마감된 대회 수를 반환하는 API
-        """
-        problems = Problem.objects.all().filter(visible=True)
-
-        # 총 문제 수
-        total_problem_length = problems.count()
-
-        # 한번이라도 accept가 된 문제 수
-        accepted_problem_length = problems.filter(accepted_number__lt=0).count()
-
-        # 마감된 대회 수
-        contests = Contest.objects.select_related("created_by").filter(visible=True)
-        cur = now()
-        contests = contests.filter(end_time__lt=cur)
-        ended_contest_length = contests.count()
-
-        home_statistics = {
-            "total_problem_length": total_problem_length,
-            "accepted_problem_length": accepted_problem_length,
-            "ended_contest_length": ended_contest_length,
-        }
-
-        return self.success(HomeStatistics(home_statistics).data)
 
 
 class UserProfileAPI(APIView):
